requirements.txt
websockets
python-dotenv
discord
protobuf
httpx
cryptography
web3
eth-abi
ujson
